# services/bot-mgmt-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing smishing analysis bots.",
        "title": "Bot Management Server API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/analyze": {
            "post": {
                "description": "Initiates a new smishing analysis task for a given URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Create a new analysis task",
                "parameters": [
                    {
                        "description": "Create Task Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/domain.AnalysisTask"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/status/{id}": {
            "get": {
                "description": "Retrieve the current status of an analysis task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Get task status",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.AnalysisTask"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/webhook": {
            "post": {
                "description": "Update task status via webhook (Internal use)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Handle webhook update",
                "parameters": [
                    {
                        "description": "Webhook Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.WebhookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.AnalysisTask": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "external_id": {
                    "description": "AWS Task ARN or similar",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "request_uuid": {
                    "description": "External User/Request UUID",
                    "type": "string"
                },
                "result": {
                    "type": "string"
                },
                "retry_count": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/domain.TaskStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "domain.TaskStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "RUNNING",
                "COMPLETED",
                "FAILED"
            ],
            "x-enum-varnames": [
                "TaskStatusPending",
                "TaskStatusRunning",
                "TaskStatusCompleted",
                "TaskStatusFailed"
            ]
        },
        "http.CreateTaskRequest": {
            "type": "object",
            "properties": {
                "request_uuid": {
                    "description": "Can be JWT or simple UUID",
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "http.WebhookRequest": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.TaskStatus"
                },
                "task_id": {
                    "description": "Matches our internal ID",
                    "type": "string"
                }
            }
        }
    }
}